backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
TensorFlow==2.13.1
Keras==2.13.1
Pillow==10.0.0
numpy==1.24.3
Werkzeug==2.3.7
